+++ spec/payload_spec.rb
@@ -1,7 +1,15 @@
-require File.join(File.dirname(File.expand_path(__FILE__)), 'base')
+require File.join( File.dirname(File.expand_path(__FILE__)), 'spec_helper')
 
 describe RestClient::Payload do
+  let(:sample_jpg_file) do
+    File.new(File.join(File.dirname(File.expand_path(__FILE__)), 'fixtures', 'master_shake.jpg'))
+  end
+  let(:sample_jpg_contents) do
+    IO.respond_to?(:binread) ? IO.binread(sample_jpg_file.path) : IO.read(sample_jpg_file.path)
+  end
+
   context "A regular Payload" do
+
     it "should use standard enctype as default content-type" do
       RestClient::Payload::UrlEncoded.new({}).headers['Content-Type'].
           should == 'application/x-www-form-urlencoded'
@@ -54,7 +62,7 @@ describe RestClient::Payload do
       RestClient::Payload::UrlEncoded.new({:foo => ['bar', 'baz']}).to_s.
           should == "foo[]=bar&foo[]=baz"
     end
-    
+
     it 'should not close if stream already closed' do
       p = RestClient::Payload::UrlEncoded.new({'foo ' => 'bar'})
       3.times {p.close}
@@ -68,9 +76,9 @@ describe RestClient::Payload do
       m.stub!(:boundary).and_return(123)
       m.headers['Content-Type'].should == 'multipart/form-data; boundary=123'
     end
-    
+
     it 'should not error on close if stream already closed' do
-      m = RestClient::Payload::Multipart.new(:file => File.new(File.join(File.dirname(File.expand_path(__FILE__)), 'master_shake.jpg')))
+      m = RestClient::Payload::Multipart.new(:file => sample_jpg_file)
       3.times {m.close}
     end
 
@@ -101,33 +109,33 @@ baz\r
     end
 
     it "should form properly separated multipart data" do
-      f = File.new(File.dirname(__FILE__) + "/master_shake.jpg")
+      f = sample_jpg_file
       m = RestClient::Payload::Multipart.new({:foo => f})
       m.to_s.should == <<-EOS
 --#{m.boundary}\r
 Content-Disposition: form-data; name="foo"; filename="master_shake.jpg"\r
 Content-Type: image/jpeg\r
 \r
-#{IO.read(f.path)}\r
+#{sample_jpg_contents}\r
 --#{m.boundary}--\r
       EOS
     end
 
     it "should ignore the name attribute when it's not set" do
-      f = File.new(File.dirname(__FILE__) + "/master_shake.jpg")
+      f = sample_jpg_file
       m = RestClient::Payload::Multipart.new({nil => f})
       m.to_s.should == <<-EOS
 --#{m.boundary}\r
 Content-Disposition: form-data; filename="master_shake.jpg"\r
 Content-Type: image/jpeg\r
 \r
-#{IO.read(f.path)}\r
+#{sample_jpg_contents.to_s}\r
 --#{m.boundary}--\r
       EOS
     end
 
     it "should detect optional (original) content type and filename" do
-      f = File.new(File.dirname(__FILE__) + "/master_shake.jpg")
+      f = sample_jpg_file
       f.instance_eval "def content_type; 'text/plain'; end"
       f.instance_eval "def original_filename; 'foo.txt'; end"
       m = RestClient::Payload::Multipart.new({:foo => f})
@@ -136,7 +144,7 @@ Content-Type: image/jpeg\r
 Content-Disposition: form-data; name="foo"; filename="foo.txt"\r
 Content-Type: text/plain\r
 \r
-#{IO.read(f.path)}\r
+#{sample_jpg_contents}\r
 --#{m.boundary}--\r
       EOS
     end
@@ -151,7 +159,7 @@ foo\r
 --#{m.boundary}--\r
       EOS
 
-      f = File.new(File.dirname(__FILE__) + "/master_shake.jpg")
+      f = sample_jpg_file
       f.instance_eval "def content_type; 'text/plain'; end"
       f.instance_eval "def original_filename; 'foo.txt'; end"
       m = RestClient::Payload::Multipart.new({:foo => {:bar => f}})
@@ -160,7 +168,7 @@ foo\r
 Content-Disposition: form-data; name="foo[bar]"; filename="foo.txt"\r
 Content-Type: text/plain\r
 \r
-#{IO.read(f.path)}\r
+#{sample_jpg_contents}\r
 --#{m.boundary}--\r
       EOS
     end
@@ -169,7 +177,7 @@ Content-Type: text/plain\r
 
   context "streamed payloads" do
     it "should properly determine the size of file payloads" do
-      f = File.new(File.dirname(__FILE__) + "/master_shake.jpg")
+      f = sample_jpg_file
       payload = RestClient::Payload.generate(f)
       payload.size.should == 22_545
       payload.length.should == 22_545
@@ -200,7 +208,7 @@ Content-Type: text/plain\r
     end
 
     it "should recognize multipart params" do
-      f = File.new(File.dirname(__FILE__) + "/master_shake.jpg")
+      f = sample_jpg_file
       RestClient::Payload.generate({"foo" => f}).should be_kind_of(RestClient::Payload::Multipart)
     end
 
@@ -213,17 +221,17 @@ Content-Type: text/plain\r
     end
 
     it "should recognize nested multipart payloads in hashes" do
-      f = File.new(File.dirname(__FILE__) + "/master_shake.jpg")
+      f = sample_jpg_file
       RestClient::Payload.generate({"foo" => {"file" => f}}).should be_kind_of(RestClient::Payload::Multipart)
     end
 
     it "should recognize nested multipart payloads in arrays" do
-      f = File.new(File.dirname(__FILE__) + "/master_shake.jpg")
+      f = sample_jpg_file
       RestClient::Payload.generate({"foo" => [f]}).should be_kind_of(RestClient::Payload::Multipart)
     end
 
     it "should recognize file payloads that can be streamed" do
-      f = File.new(File.dirname(__FILE__) + "/master_shake.jpg")
+      f = sample_jpg_file
       RestClient::Payload.generate(f).should be_kind_of(RestClient::Payload::Streamed)
     end